requirements.txt
opencv-python>=4.10
numpy>=1.26
pyserial>=3.5
